requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
djangorestframework==3.15.0